+++ server/database/migrations/09 - create_tags_questions.js
@@ -1,10 +1,14 @@
 exports.up = async function up(knex){
-    return knex.schema.createTable('tags_questions', table => {
-        table.increments('id').primary(),
-        table.integer('tag_id').unsigned().notNullable().references('id').inTable('tags').onDelete('CASCADE').index();
-        table.integer('question_id').unsigned().notNullable().references('id').inTable('questions').onDelete('CASCADE').index();
-    })
 
+    knex.schema.hasTable('tags_questions').then(function(exists){
+        if (!exists) {
+            return knex.schema.createTable('tags_questions', table => {
+                table.increments('id').primary(),
+                table.integer('tag_id').unsigned().notNullable().references('id').inTable('tags').onDelete('CASCADE').index();
+                table.integer('question_id').unsigned().notNullable().references('id').inTable('questions').onDelete('CASCADE').index();
+            })
+        }
+    })
 }
